Refactor emoji and replace compatitable IE icon
